requirements.txt
streamlit
requests
python-dotenv
pymongo
